==> 2_Letsgo_preprocess.ipynb <==
# Code 5-16 to Code 5-37

For this exercise we are going to take a dialog dataset “Lets Go” bus dataset. The details of the dataset can be found https://dialrc.github.io/LetsGoDataset/ (Raux et al. 2005)Let’s Go was funded by the National Science Foundation. Have cleaned the dataset to represent a conversation format as show in "Clean_lets_go.ipynb"

import pandas as pd
t2 = pd.read_csv("letsgo_op.csv")

t2.head()

len(t2)

t2.loc[t2.cust_text.isna()==True,"cust_text"]="None"

Given this data is about bus schedule customers would often use bus stop and place names. We took bus stop names of Pittsburg and added few more names to the list like - downtown, malls, college etc. We would use this file to replace some of the bus stop names with some common name to help in genearlization

import numpy as np
bus_sch = np.load("bus_stops.npz",allow_pickle=True) #make sure you use the .npz!
bus_sch1 = list(bus_sch['arr_0'])

bus_sch1

Replacing some common sentences in the bot_text

string_repl = "welcome to the cmu let's go bus information system.  to get help at any time, just say help or press zero. what can i do for you?  i am an automated spoken dialogue system that can give you schedule information for bus routes in pittsburgh's east end. you can ask me about the following buses: 28x, 54c, 56u, 59u, 61a, 61b, 61c, 61d, 61f, 64a, 69a, and 501. what bus schedule information are you looking for? for example, you can say, when is the next 28x from downtown to the airport? or i'd like to go from mckeesport to homestead tomorrow at 10 a.m."
t2["bot_text"] = t2["bot_text"].str.replace("welcome.*mckeesport to homestead.*automated spoken dialogue.*","greeting_templ")
t2["trim_bot_txt"] = t2["bot_text"].str.replace('goodbye.*','goodbye')


Now we will replace the place names in cust_text and bot_text using the bus_sch1 list with the function in Code 19 -  repl_place_bus using “fuzz.ratio” between words in the sentences and bus_sch1.”fuzz.ratio” calculates the similarity between strings using Levenshtein Distance (minimum number of single character edits required to change one word to another). “fuzzywuzzy” is the package used to calculate “fuzz.ratio”. You can “pip” install “fuzzywuzzy”


from fuzzywuzzy import fuzz 
str_list = []
def repl_place_bus(row,req_col):
    #for i,row in t2.iterrows():
    ctext = row[req_col]
    fg=0
    str1=ctext

    try:
        for j in ctext.split():

            for k in bus_sch1:
                score = fuzz.ratio(j,k )
                if(score>=70):
                    #print (i,j,k)
                    fg=1
                    break;
            #if(fg==1):
                #break;
            if(fg==1):
                fg=0
                str1 = str1.replace(j," place_name ")

    except:
        print (j,i)
    #str_list.append(str1)
        
    return str1

We have 7000+ rows. In each row we are iterating through the words in the sentences. This would increase the run-time very much. Inorder to run it faster, we will be applying the functions on parallelization in python. We would run the function simultanelusly across multiple “processes” using a library called “joblib”. The “n_jobs” parameter provides the number of processes that are going to be run in parallel.

Applying the function for customer text

from joblib import Parallel, delayed
com0 = Parallel(n_jobs=12, backend="threading",verbose=1)(delayed(repl_place_bus)(row,"cust_text") for i,row in t2.iterrows())

Applying the function for bot text

bot_list = Parallel(n_jobs=12, backend="threading",verbose=1)(delayed(repl_place_bus)(row,"trim_bot_txt") for i,row in t2.iterrows())

len(com0),len(t2),len(bot_list)

t2.head()

t2["corrected_cust"] = com0
t2["corrected_bot"] = bot_list

The columns “corrected_cust” and “corrected_bot” contains the replaced sentences. We will now do the following preprocesses - directions (since its a bus booking), normalizing am/pm, replacing number words to numbers, bus names (like 8a,11c etc), time, replacing numbers

direct_list = ['east','west','north','south']

def am_pm_direct_repl(t2,col):
    t2[col] = t2[col].str.replace('p.m','pm')
    t2[col] = t2[col].str.replace('p m','pm')
    t2[col] = t2[col].str.replace('a.m','am')
    t2[col] = t2[col].str.replace('a m','am')
    
    for i in direct_list:
        t2[col] = t2[col].str.replace(i,' direction ')
   
    return t2

t3 =  am_pm_direct_repl(t2,"corrected_cust")
t3 =  am_pm_direct_repl(t2,"corrected_bot")

The following function using “num2words” package gets numeric numbers for number-words like “one”, “two”, “three” etc. We then use this to replace our sentences with the right numbers. We would be replacing “th” words. For eg -fifth avenue gets replaced with “num_th”.

from num2words import num2words
word_num_list = []
word_num_th_list = []
for i in range(0,100):
    word_num_list.append(num2words(i))
    word_num_th_list.append(num2words(i,to='ordinal'))

print (word_num_list[0:5])
print (word_num_th_list[0:5])

def repl_num_words(t2,col):
    for num,i in enumerate(word_num_th_list):
        t2[col] = t2[col].str.replace(i,"num_th")
        t2[col] = t2[col].str.replace(word_num_list[num],str(num))
    return t2

t4 =  repl_num_words(t3,"corrected_cust")
t4 =  repl_num_words(t3,"corrected_bot")

Replacing bus and time names herewith

t4["corrected_cust"] = t4["corrected_cust"].str.replace('[0-9]{1,2}[a-z]{1,2}','bus_name')
t4["corrected_cust"] = t4["corrected_cust"].str.replace('[0-9]{1,2}[\.\s]*[0-9]{1,2}[\s]*pm','time_name')
t4["corrected_cust"] = t4["corrected_cust"].str.replace('[0-9]{1,2}[\.\s]*[0-9]{1,2}[\s]*am','time_name')

t4["corrected_bot"] = t4["corrected_bot"].str.replace('[0-9]{1,2}[a-z]{1,2}','bus_name')
t4["corrected_bot"] = t4["corrected_bot"].str.replace('[0-9]{1,2}[\.\s]*[0-9]{0,2}[\s]*pm','time_name')
t4["corrected_bot"] = t4["corrected_bot"].str.replace('[0-9]{1,2}[\.\s]*[0-9]{0,2}[\s]*am','time_name')

Next  we replace special characters and split the bot sentences with a sentence separator “sep_sent”. This will help us to “templatize” each of the bot sentences and generate these templates from the decoder  later

t4["corrected_bot1"] = t4["corrected_bot"].str.replace('\xa0','')
t4["corrected_bot1"] = t4["corrected_bot1"].str.replace('[^a-z0-9\s\.\?]+','')
t4["corrected_bot1"] = t4["corrected_bot1"].str.replace('[0-9]{1,3}',' short_num ')
t4["corrected_bot1"] = t4["corrected_bot1"].str.replace('[0-9]{4,}',' long_num ')

t4["corrected_bot1"] = t4["corrected_bot1"].str.replace('(\.\s){1,3}',' sep_sent ')
t4["corrected_bot1"] = t4["corrected_bot1"].str.replace('(\.){1,3}',' sep_sent ')
t4["corrected_bot1"] = t4["corrected_bot1"].str.replace('\?',' sep_sent ')
t4["corrected_bot1"] = t4["corrected_bot1"].str.replace('(sep_sent )+',' sep_sent ')



Replacing special characters in customer sentences. This is done in order to decrease the sparsity of the data. For example  - consider 2 sentences - “is this correct...” and “is this correct”. The word “correct” is same except in both cases – except one has 3 dots and the other has none. When we tokenize we want to make sure that these 2 words are treated the same 

t4["corrected_cust"] = t4["corrected_cust"].str.replace('(\.\s){1,3}',' ')
t4["corrected_cust"] = t4["corrected_cust"].str.replace('(\.){1,3}',' ')

t4["corrected_cust"].to_clipboard()


len(t3["corrected_bot"].value_counts())

Since we have taken a corpus which is an interaction between a bot and customer, some of the sentences are repetitive across different conversations. Given the nature of customer service data even if it had been a human to human conversation, there would be a good amount of repetitive sentences used across different dialogues. These repetitive sentences can be converted to “templates”. 

templ = pd.DataFrame(t4["corrected_bot1"].value_counts()).reset_index()
templ.columns = ["sents","count"]

templ.head()

Getting all the unique sentences into a list. Remaking the list into a dataframe to assign each sentence to a template id

df = templ.sents.str.split("sep_sent",expand=True)
sents_all = []
for i in df.columns:
    l1 = list(df[i].unique())
    sents_all = sents_all + l1
sents_all1 = set(sents_all)

sents_all_df = pd.DataFrame(sents_all1)
sents_all_df.columns = ["sent"]
sents_all_df["ind"] = sents_all_df.index
sents_all_df["ind"] = "templ_" + sents_all_df["ind"].astype('str')

sents_all_df.head()

sents_all_df.to_clipboard()

sents_all_df["cnt"]  = sents_all_df["sent"].str.count(" ")

sents_all_df["cnt"].describe()

Now we take our original dataset of t4 and assign different sentences to the template ids. First we create a dictionary with sentences as the keys and template ids

sent_all_df1 = sents_all_df.loc[:,["sent","ind"]]
df_sents = sent_all_df1.set_index('sent').T.to_dict('list')

t4["corrected_bots_sents"] = t4["corrected_bot1"].str.split("sep_sent")

index_list=[]
for i, row in t4.iterrows():
    sent_list = row["corrected_bots_sents"]
    str_index = ""
    for j in sent_list:
        if(len(j)>=3):
            str_index = str_index + " "  + str(df_sents[j][0])
    index_list.append(str_index)
    

t4["bots_templ_list"] = index_list

t4.tail()

Our training set is like bot asks a question and the customer says the answer. But for training our input would be customer sentence and the response from the bot as the output. So we will shift our “corrected_bot1” to one sentence after. That way the corpus becomes incoming customer text and the right output from the bot. We are shifting the user_id to mark the end of  the chat

t4["u_id_shift"] = t4["user_id"].shift(-1)
t4["corrected_bot1_shift"] = t4["corrected_bot1"].shift(-1)
t4["bots_templ_list_shift"] = t4["bots_templ_list"].shift(-1)

t4.loc[t4.u_id_shift!=t3.user_id,"corrected_bot1_shift"] = "end_of_chat"
t4.loc[t4.u_id_shift!=t3.user_id,"bots_templ_list_shift"] = "end_of_chat"

req_cols = ["user_id","corrected_cust","corrected_bot1_shift","bots_templ_list_shift"]
t5 = t4[req_cols]

t5.tail()

Save the processed dataset

t5.to_csv("lets_go_model_set1.csv")

We are now pickling the output of dictionary “df_sents” for mapping the templates to original sentences. In the encoder we would use “template_ids” for output. Then we remap using the dictionary to prin the actual sentences

import pickle
output = open('dict_templ.pkl', 'wb')

# Pickle dictionary using protocol 0.
pickle.dump(df_sents, output)
