==> 5_Ktrain_bert_v1.ipynb <==
# Code 5-66 to Code 5-72

This code uses "google colab" for the BERT model through Ktrain (ktrain: A Low-Code Library for Augmented Machine Learning, Arun S Maiya https://arxiv.org/abs/2004.10703) 

!pip install tensorflow-gpu==2.0.0
!pip install ktrain

Upload file of Bank Classification - bank4.csv

from google.colab import drive
drive.mount('/content/drive/')
import pandas as pd
root_dir = "/content/drive/My Drive/Newbounty_resync/Apress/Chapters/Chapter5"

base_fl_csv =  root_dir + '/bank4.csv'
t1 = pd.read_csv(base_fl_csv,encoding='latin1')

We are using TPU instance. This code checks the configuration


# import os
# import pprint
# import tensorflow as tf

# if 'COLAB_TPU_ADDR' not in os.environ:
#   print('ERROR: Not connected to a TPU runtime; please see the first cell in this notebook for instructions!')
# else:
#   tpu_address = 'grpc://' + os.environ['COLAB_TPU_ADDR']
#   print ('TPU address is', tpu_address)

#   with tf.session(tpu_address) as session:
#     devices = session.list_devices()
    
#   print('TPU devices:')
#   pprint.pprint(devices)


t1.head()

Splitting data in train and validation using stratified sampling

tgt = t1["Final intent"]

from sklearn.model_selection import StratifiedShuffleSplit
sss = StratifiedShuffleSplit(test_size=0.5,random_state=42,n_splits=1)

for train_index, test_index in sss.split(t1, tgt):
    x_train, x_test = t1[t1.index.isin(train_index)], t1[t1.index.isin(test_index)]
    y_train, y_test = t1.loc[t1.index.isin(train_index),"Final intent"], t1.loc[t1.index.isin(test_index),"Final intent"]

Converting the dependent variables into lists



y_train1 = y_train.values.tolist()
y_test1 = y_test.values.tolist()


type(y_train1)

class1 = list(t1["Final intent"].unique())

class1

type(y_train)

This dataset is converted to BERT specific format in below code. This code is referred from  https://www.kaggle.com/ksaxena10/bert-sentiment-analysis 

import ktrain 
from ktrain import text
import numpy as np
(x_train2,  y_train2), (x_test2, y_test2), preproc = text.texts_from_array(x_train=np.array(x_train["Line"]), y_train=y_train1,
                                                                       x_test=np.array(x_test["Line"]), y_test=y_test1,
                                                                       class_names=class1,
                                                                       preprocess_mode='bert',
                                                                       ngram_range=1, 
                                                                       maxlen=350)

The data is tokenized using BERT tokenization and numbers assigned for the given sequence length. Using preproc = "BERT" we tell the format that is needed to be kept

x_train2[0].shape

x_train2[0][1]

Building the BERT model

We pass 'bert' as a parameter to "text_classifier" so that the BERT model is fine tuned. The preproc parameter auto generates the format that "BERT" requires

model = text.text_classifier('bert', train_data=(x_train2, y_train2), preproc=preproc)
learner = ktrain.get_learner(model, train_data=(x_train2, y_train2), batch_size=6)

"2e-5" is the reccomended learning rate

hist = learner.fit_onecycle(2e-5, 2) 

predictor = ktrain.get_predictor(learner.model, preproc)

predictor.get_classes()

Executing on a sample line

predictor.predict(['waive my annual fee'])
